Move MMR diversification to its own file
