+++ src/poprox_recommender/default.py
@@ -14,6 +14,7 @@ from tqdm import tqdm
 from transformers import AutoTokenizer
 
 from poprox_concepts import Article, ClickHistory, InterestProfile
+from poprox_recommender.diversifiers.mmr import mmr_diversification
 from poprox_recommender.model.nrms import NRMS
 from poprox_recommender.paths import model_file_path
 from poprox_recommender.topics import normalized_topic_count, user_topic_preference
@@ -112,38 +113,6 @@ def build_user_embedding(click_history: ClickHistory, article_embeddings, model,
     )
 
     return model.get_user_vector(clicked_news_vector)
-
-
-def mmr_diversification(rewards, similarity_matrix, theta: float, topk: int):
-    # MR_i = \theta * reward_i - (1 - \theta)*max_{j \in S} sim(i, j) # S us
-    # R is all candidates (not selected yet)
-
-    S = []  # final recommendation (topk index)
-    # first recommended item
-    S.append(rewards.argmax())
-
-    for k in range(topk - 1):
-        candidate = None  # next item
-        best_MR = float("-inf")
-
-        for i, reward_i in enumerate(rewards):  # iterate R for next item
-            if i in S:
-                continue
-            max_sim = float("-inf")
-
-            for j in S:
-                sim = similarity_matrix[i][j]
-                if sim > max_sim:
-                    max_sim = sim
-
-            mr_i = theta * reward_i - (1 - theta) * max_sim
-            if mr_i > best_MR:
-                best_MR = mr_i
-                candidate = i
-
-        if candidate is not None:
-            S.append(candidate)
-    return S
 
 
 def pfar_diversification(relevance_scores, articles, topic_preferences, lamb, tau, topk):
